refactor: extract config parsers from discovery/__init__.py into discovery/config_parsers.py
Move all parse_*, audit_*, and _parse_docker_mcp_catalog functions (602 lines)
into a dedicated config_parsers module. discovery/__init__.py drops from
1,750 → 1,204 lines (-31%). All symbols re-exported for backward compatibility.
